requirements.txt
aiogram
aiohttp
apscheduler
cryptography
beautifulsoup4
python-dotenv
psycopg2-binary
pyOpenSSL
whois
ipwhois

